Tweaked $::osfamily calls in init.pp and params
facter osfamily has no value on SLES 11 SP3 during my tests,
consequently SLES is not recognized as belonging to
the Suse osfamily. Then I added a test in params.pp to give a
value to $snmp::params::osfamily which is 'Suse' if $::operatingsystem
is 'SLES' or $::osfamily if not. I also changed calls for $::osfamily
to $snmp::params::osfamily. Maybe this problem is due to an inconsistent
reason or because my tests are made on virtual machines (??). For now
this does the trick for me.