Compare existing and expected passwords to trigger user creation
This rewrite of `snmp::snmpv3_user` (re)creates the user if it does not
exist or the stored passwords do not match the expected values. It uses the
`snmpv3_user` fact to retrieve the currently configured SNMPv3 users and
calculates the expected password hashes by calling the
`snmp::snmpv3_usm_hash` function using the cleartext password.

Updating the passwords in Puppet will now correctly update the passwords in
the SNMP configuration file. (closes #4).