Set 'var-net-snmp' ensure to absent when removing
Previously 'File[var-net-snmp]' was always getting set to 'directory',
even when trying to remove snmp from a system. If the owner or group
used for 'var-net-snmp' were removed before this, the resource would
raise an error due to them missing, but if the ensure is set to absent
it will not check for user/group existance.